far->FAR

+++ src/eg3dcam.c
@@ -16,7 +16,7 @@ void setViewRotation(int rotX, int rotY, int rotZ) {
 }
 
 // ==== seg000:0x3a90 ====
-void setViewPosition(int viewX, int viewY, int viewZ) { /* Original: SetViewPos(X,Y,Z). Store viewer coordinates for 3D transforms. */
+void setViewPosition(int16 viewX, int16 viewY, int16 viewZ) { /* Original: SetViewPos(X,Y,Z). Store viewer coordinates for 3D transforms. */
     g_viewPosX = viewX;
     g_viewPosY = viewY;
     g_viewPosZ = viewZ;
