+++ src/eg3dcam.h
@@ -2,8 +2,8 @@
 #define F15_SE2_EG3DCAM
 /* public interface of eg3dcam.c */
 
-void setViewRotation(int, int, int);
-void setViewPosition(int viewX, int viewY, int viewZ);
+void setViewRotation(int16, int16, int16);
+void setViewPosition(int16 viewX, int16 viewY, int16 viewZ);
 void setViewPositionFrac(int fracX, int fracY, int fracZ);
 
 #endif /* F15_SE2_EG3DCAM */
